chore #37: 동네 삭제 User -> UserLocation 패키지 이동
