+++ be/carrot/src/main/java/com/example/carrot/user/service/UserService.java
@@ -155,27 +155,4 @@ public class UserService {
 		userRepository.updateRefreshTokenByUserIdAndRefreshToken(userId, logoutRequestDto.getRefreshToken());
 	}
 
-	@Transactional
-	public UserLocationDeleteResponseDto deleteUserLocation(Long locationId, Long userId) {
-		User user = userRepository.findByUserId(userId)
-			.orElseThrow(() -> new CustomException(StatusCode.NOT_FOUND_USER));
-		Location location = locationRepository.findByLocationId(locationId)
-			.orElseThrow(() -> new CustomException(StatusCode.NOT_FOUND_LOCATION));
-
-		// 등록된 동네가 하나라면 제거 불가능 -> 예외 처리
-		if (isOneLocation(user)) {
-			throw new CustomException(StatusCode.DELETE_LOCATION_EXCEPTION);
-		}
-
-		// 등록된 동네가 두개라면 제거 가능 -> 제거 후 남은 동네를 mainLocation으로 변경
-		UserLocation deletedUserLocation = user.deleteUserLocation(location);
-		userLocationRepository.delete(deletedUserLocation);
-
-		return UserLocationDeleteResponseDto.of(user.findMainLocation());
-	}
-
-	private boolean isOneLocation(User user) {
-		final int ONE = 1;
-		return user.getUserLocations().size() == ONE;
-	}
 }
