+++ be/carrot/src/main/java/com/example/carrot/user_location/controller/UserLocationController.java
@@ -4,8 +4,10 @@ import java.util.List;
 
 import javax.validation.Valid;
 
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.example.carrot.global.common.ApiResponse;
 import com.example.carrot.location.dto.request.MainLocationRequestDto;
 import com.example.carrot.location.dto.response.MainLocationResponseDto;
+import com.example.carrot.user.dto.response.UserLocationDeleteResponseDto;
 import com.example.carrot.user_location.dto.response.ReadUserLocationResponseDto;
 import com.example.carrot.user_location.service.UserLocationService;
 
@@ -44,5 +47,11 @@ public class UserLocationController {
 		MainLocationResponseDto mainLocationResponseDto = userLocationService.registerMainLocation(
 			mainLocationRequestDto, userId);
 		return ApiResponse.success(mainLocationResponseDto);
+	}
+
+	@DeleteMapping("/users/locations/{locationId}")
+	public ApiResponse<UserLocationDeleteResponseDto> deleteUserLocation(@PathVariable Long locationId,
+		@RequestAttribute Long userId) {
+		return ApiResponse.success(userLocationService.deleteUserLocation(locationId, userId));
 	}
 }
