+++ be/carrot/src/main/java/com/example/carrot/user_location/service/UserLocationService.java
@@ -14,6 +14,7 @@ import com.example.carrot.location.dto.response.MainLocationResponseDto;
 import com.example.carrot.location.entity.Location;
 import com.example.carrot.location.repository.LocationRepository;
 import com.example.carrot.location.service.LocationService;
+import com.example.carrot.user.dto.response.UserLocationDeleteResponseDto;
 import com.example.carrot.user.entity.User;
 import com.example.carrot.user.repository.UserRepository;
 import com.example.carrot.user_location.dto.response.ReadUserLocationResponseDto;
@@ -96,6 +97,30 @@ public class UserLocationService {
 		}
 
 		return readUserLocationResponseDtos;
+	}
+
+	@Transactional
+	public UserLocationDeleteResponseDto deleteUserLocation(Long locationId, Long userId) {
+		User user = userRepository.findByUserId(userId)
+			.orElseThrow(() -> new CustomException(StatusCode.NOT_FOUND_USER));
+		Location location = locationRepository.findByLocationId(locationId)
+			.orElseThrow(() -> new CustomException(StatusCode.NOT_FOUND_LOCATION));
+
+		// 등록된 동네가 하나라면 제거 불가능 -> 예외 처리
+		if (isOneLocation(user)) {
+			throw new CustomException(StatusCode.DELETE_LOCATION_EXCEPTION);
+		}
+
+		// 등록된 동네가 두개라면 제거 가능 -> 제거 후 남은 동네를 mainLocation으로 변경
+		UserLocation deletedUserLocation = user.deleteUserLocation(location);
+		userLocationRepository.delete(deletedUserLocation);
+
+		return UserLocationDeleteResponseDto.of(user.findMainLocation());
+	}
+
+	private boolean isOneLocation(User user) {
+		final int ONE = 1;
+		return user.getUserLocations().size() == ONE;
 	}
 
 }
